Rename Client as SocketClient, remove command features
