--- lib/log.js
@@ -1,5 +1,0 @@
-var bunyan = require('bunyan');
-module.exorts = bunyan.createLogger({
-  name: 'biodome-client',
-  stream: process.stdout
-});
