+++ lib/socket-client.js
@@ -1,33 +1,29 @@
 var ws = require('ws')
   , util = require('util')
-  , EventEmitter = require('events').EventEmitter
-  , log = require('./log')
-  , messageValidator = require('./message-validator').factory();
+  , EventEmitter = require('events').EventEmitter;
 
-function Client(conf)
-{
+function SocketClient(conf) {
   // { host : str, ssl : bool }
   this.conf = conf;
   EventEmitter.call(this);
   var protocol = (this.conf.ssl) ? 'wss://' : 'ws://';
   // TODO reconnection strategy
   this.connection = new ws(protocol + this.conf.host);
-  this.messageValidator = messageValidator;
   this.wireEvents();
 }
 
-util.inherits(Client, EventEmitter);
+util.inherits(SocketClient, EventEmitter);
 
-var proto = Client.prototype;
+var proto = SocketClient.prototype;
 
-proto.wireEvents = function()
-{
+proto.wireEvents = function() {
   var self = this;
   this.connection.on('open', this.socketOpen.bind(this));
   this.connection.on('close', this.socketClose.bind(this));
   this.connection.on('message', function(message) {
     self.socketMessage(message);
   });
+
   this.connection.on('error', function(error) {
     self.socketError(error);
   });
@@ -48,38 +44,18 @@ proto.socketMessage = function(msg) {
     return this.socketError('Server returned invalid response');
   }
 
-  var meta = messageValidator.parse(data);
-  if (meta.valid) {
-    if ('error' == data.type) {
-      return this.messageError(data.message);
-    }
-    this.emit(data.type + ':update', data);
-  } else {
-    return this.socketError('Server sent invalid message: ');
-  }
-};
-
-proto.messageError = function(error) {
-  //log('error', 'message error: ' + error);
-  this.emit('message_error', error);
+  this.emit('data', data);
 };
 
 proto.socketError = function(error) {
-  //log('error', error);
   this.emit('server_error', error);;
 };
 
-proto.setDeviceState = function(deviceId, state)
-{
-  this.connection.send('device:' + deviceId + ':' + state);
-};
-
-proto.close = function()
-{
+proto.close = function() {
   this.connection.close();
 };
 
-module.exports = Client;
-module.exports.factory = function(conf) {
-  return new Client(conf);
+module.exports = SocketClient;
+module.exports.new = function(conf) {
+  return new SocketClient(conf);
 }
